Added cmd= and updated news item

+++ R/fread.R
@@ -1,6 +1,8 @@
 
-fread <- function(input="",file=NULL,text=NULL,sep="auto",sep2="auto",dec=".",quote="\"",nrows=Inf,header="auto",na.strings=getOption("datatable.na.strings","NA"),stringsAsFactors=FALSE,verbose=getOption("datatable.verbose",FALSE),skip="__auto__",select=NULL,drop=NULL,colClasses=NULL,integer64=getOption("datatable.integer64","integer64"), col.names, check.names=FALSE, encoding="unknown", strip.white=TRUE, fill=FALSE, blank.lines.skip=FALSE, key=NULL, index=NULL, showProgress=getOption("datatable.showProgress",interactive()), data.table=getOption("datatable.fread.datatable",TRUE), nThread=getDTthreads(), logical01=getOption("datatable.logical01", FALSE), autostart=NA)
+fread <- function(input="",file=NULL,text=NULL,cmd=NULL,sep="auto",sep2="auto",dec=".",quote="\"",nrows=Inf,header="auto",na.strings=getOption("datatable.na.strings","NA"),stringsAsFactors=FALSE,verbose=getOption("datatable.verbose",FALSE),skip="__auto__",select=NULL,drop=NULL,colClasses=NULL,integer64=getOption("datatable.integer64","integer64"), col.names, check.names=FALSE, encoding="unknown", strip.white=TRUE, fill=FALSE, blank.lines.skip=FALSE, key=NULL, index=NULL, showProgress=getOption("datatable.showProgress",interactive()), data.table=getOption("datatable.fread.datatable",TRUE), nThread=getDTthreads(), logical01=getOption("datatable.logical01", FALSE), autostart=NA)
 {
+  if (missing(input)+is.null(file)+is.null(text)+is.null(cmd) < 3L) stop("Used more than one of the arguments input=, file=, text= and cmd=.")
+  input_literal = identical(class(substitute(input)), "character")  # to allow commands in input= iff literal character constant (not variable, not expression). See news v1.11.6.
   if (is.null(sep)) sep="\n"         # C level knows that \n means \r\n on Windows, for example
   else {
     stopifnot( length(sep)==1L, !is.na(sep), is.character(sep) )
@@ -25,23 +27,18 @@ fread <- function(input="",file=NULL,text=NULL,sep="auto",sep2="auto",dec=".",qu
   nThread=as.integer(nThread)
   stopifnot(nThread>=1L)
   if (!is.null(text)) {
-    if (!identical(input, "")) stop("You can provide 'input=' or 'text=', not both.")
-    if (!is.null(file)) stop("You can provide 'file=' or 'text=', not both.")
     if (!is.character(text)) stop("'text=' is type ", typeof(text), " but must be character.")
     if (!length(text)) return(data.table())
     if (length(text) > 1L) {
-      tmpFile = tempfile()
-      on.exit(unlink(tmpFile), add=TRUE)
-      cat(text, file=tmpFile, sep="\n")  # avoid paste0() which could create a new very long single string in R's memory
+      cat(text, file=(tmpFile<-tempfile()), sep="\n")  # avoid paste0() which could create a new very long single string in R's memory
       input = tmpFile
+      on.exit(unlink(tmpFile), add=TRUE)
     } else {
       # avoid creating a tempfile() for single strings, which can be done a lot; e.g. in the test suite.
       input = text
     }
-    text = NULL
   }
   else if (!is.null(file)) {
-    if (!identical(input, "")) stop("You can provide 'input=' or 'file=', not both.")
     file_info = file.info(file)
     if (is.na(file_info$size)) stop("File '",file,"' does not exist or is non-readable.")
     if (isTRUE(file_info$isdir)) stop("File '",file,"' is a directory. Not yet implemented.") # dir.exists() requires R v3.2+, #989
@@ -50,11 +47,12 @@ fread <- function(input="",file=NULL,text=NULL,sep="auto",sep2="auto",dec=".",qu
       return(if (data.table) data.table(NULL) else data.frame(NULL))
     }
     input = file
-  } else {
+  }
+  else if (is.null(cmd)) {
     if (!is.character(input) || length(input)!=1L) {
-      stop("'input' must be a single character string containing a file name, a system command containing at least one space, a URL starting 'http[s]://', 'ftp[s]://' or 'file://', or, the input data itself containing at least one \\n or \\r")
+      stop("input= must be a single character string containing a file name, a system command containing at least one space, a URL starting 'http[s]://', 'ftp[s]://' or 'file://', or, the input data itself containing at least one \\n or \\r")
     }
-    if ( input == "" || length(grep('\\n|\\r', input)) ) {
+    if (input=="" || length(grep('\\n|\\r', input))) {
       # input is data itself containing at least one \n or \r
     } else {
       file_info = file.info(input)
@@ -69,36 +67,43 @@ fread <- function(input="",file=NULL,text=NULL,sep="auto",sep2="auto",dec=".",qu
         if (substring(input,1L,1L)==" ") {
           stop("Input argument is not a file name and contains no \\n or \\r, but starts with a space. Please remove the leading space.")
         }
-        # either a download or a system command, both to temp file
-        tmpFile = tempfile()
-        on.exit(unlink(tmpFile), add=TRUE)
         str6 = substring(input,1L,6L)   # avoid grepl() for #2531
         str7 = substring(input,1L,7L)
         str8 = substring(input,1L,8L)
         if (str7=="ftps://" || str8=="https://") {
           if (!requireNamespace("curl", quietly = TRUE))
             stop("Input URL requires https:// connection for which fread() requires 'curl' package, but cannot be found. Please install curl using 'install.packages('curl')'.")
-          curl::curl_download(input, tmpFile, mode="wb", quiet = !showProgress)
+          curl::curl_download(input, tmpFile<-tempfile(), mode="wb", quiet = !showProgress)
+          input = tmpFile
+          on.exit(unlink(tmpFile), add=TRUE)
         }
         else if (str6=="ftp://" || str7== "http://" || str7=="file://") {
           method = if (str7=="file://") "internal" else getOption("download.file.method", default="auto")
           # force "auto" when file:// to ensure we don't use an invalid option (e.g. wget), #1668
-          download.file(input, tmpFile, method=method, mode="wb", quiet=!showProgress)
+          download.file(input, tmpFile<-tempfile(), method=method, mode="wb", quiet=!showProgress)
           # In text mode on Windows-only, R doubles up \r to make \r\r\n line endings. mode="wb" avoids that. See ?connections:"CRLF"
+          input = tmpFile
+          on.exit(unlink(tmpFile), add=TRUE)
         }
         else if (length(grep(' ', input))) {
-          (if (.Platform$OS.type == "unix") system else shell)(paste0('(', input, ') > ', tmpFile))
+          if (input_literal) cmd = input
+          else stop("input= looks like a system command but has been constructed with paste (or similar) or passed in via a variable. Feasibly, this could be an attacker passing a system command to your code (via your GUI or API) when you're expecting a file name or some literal data. For security and clarity please use cmd=, text= or file=. See NEWS file for v1.11.6 for more information.")
         }
         else stop("File '",input,"' does not exist; getwd()=='", getwd(), "'",
                   ". Include correct full path, or one or more spaces to consider the input a system command.")
-        input = tmpFile  # the file name
         if (!file.info(input)$size) {
+          # put this at C level
           warning(sprintf("File '%s' has size 0. Returning a NULL %s.",
                           input, if (data.table) 'data.table' else 'data.frame'))
           return(if (data.table) data.table(NULL) else data.frame(NULL))
         }
       }
     }
+  }
+  if (!is.null(cmd)) {
+    (if (.Platform$OS.type == "unix") system else shell)(paste0('(', cmd, ') > ', tmpFile<-tempfile()))
+    input = tmpFile
+    on.exit(unlink(tmpFile), add=TRUE)
   }
   if (!missing(autostart)) warning("'autostart' is now deprecated and ignored. Consider skip='string' or skip=n");
   if (is.logical(colClasses)) {
